Bug fixes and network/gps operation enhancements.
Fixed bug where the network would never come up if the device was turned
on in an area without signal.
Changed GPS tick code to stay on until such time as a gps lock is
received or a timeout occurs. Timeout is a simple iteration count.
Added code to set the APN based on the operator name stored on the sim
card.
Added a second thread for sending cell data in the event that the gps
has not locked for a large period of time.
Added device.lua which is aimed at handling operations at a higher layer
than the at_command and at_abs libraries.
Fixed bug in tcp client where the app handles were not being handled
correctly.
Added some example serializer that will come in handy if writing to file
needed at some point.

+++ simreporter/reporter.py
@@ -1,25 +1,45 @@
 from atlib import *
 import time
+from __builtin__ import file
 
 
 #module = serial.Serial("/dev/cu.usbserial-A105NJ7M",  115200, timeout=5)
 module = serial.serial_for_url("rfc2217://10.1.1.5:9990", 115200, timeout=5)
 # module = serial.Serial("/dev/ttyUSB0",  115200, timeout=5)
 
+import os
+
+
+def touch(fname, times=None):
+    with open(fname, 'a'):
+        os.utime(fname, times)
+
+
+import os.path
+
+
+def file_is_newer_than(file1, file2):
+    if os.path.isfile(file1) and os.path.isfile(file2):
+        return os.path.getctime(file1) > os.path.getctime(file2)
+    return True
+
+
 try:
     get_response(module)
     change_dir(module, "c:/")
     files = [
         "at_abs.lua",
-        #         "at_commands.lua",
-        #         "canary.lua",
-        #         "network_setup.lua",
-        #         "reporter.lua",
-        #         "tcp_client.lua",
-        #         "encapsulation.lua",
-        #         "nmea_getinfo.lua",
-        #         "util.lua",
+        "at_commands.lua",
+        "canary.lua",
+        "device.lua",
+        "encapsulation.lua",
+        "network_setup.lua",
+        "nmea_getinfo.lua",
+        "reporter.lua",
+        "tcp_client.lua",
+        "util.lua",
     ]
+    compile_files = []
     stop_script(module)
     while(script_is_running(module)):
         print "Script still running"
@@ -27,23 +47,31 @@ try:
     for file in files:  # os.listdir("."):
         if os.path.isfile(file):
             if "lua" in file:
-                print "Putting file " + file
-                with open(file, 'r') as content_file:
-                    content = content_file.read()
-                    put_file(module, file, content)
-                    # delete_file(module, file)
+                if file_is_newer_than(file, "lastupload/" + file):
+                    print "Putting file " + file
+                    with open(file, 'r') as content_file:
+                        content = content_file.read()
+                        put_file(module, file, content)
+                        # delete_file(module, file)
+                    touch("lastupload/" + file)
+                    compile_files.append(file)
+                else:
+                    print "File %s is not newer than last uploaded version" % (
+                        file
+                    )
 
     # for file in files:  # os.listdir("."):
     #     compiled = file.replace(".lua", ".out")
     #     delete_file(module, compiled)
 
-    for file in files:  # os.listdir("."):
+    for file in compile_files:  # os.listdir("."):
         compile_file(module, file)
 
     # for file in files:  # os.listdir("."):
     #    delete_file(module, file)
 
-    delete_file(module, "autorun.out")
+    if "nmea_getinfo.lua" in compile_files:
+        delete_file(module, "autorun.out")
 
     # compile_file(module, "reporter.lua")
     ls(module)
@@ -58,7 +86,8 @@ try:
     #    get_response(module, 2)
     #    get_response(module, 2)
     #    run_script(module, "nmea_getinfo.out")
-    copy_file(module, "nmea_getinfo.out", "autorun.out")
+    if "nmea_getinfo.lua" in compile_files:
+        copy_file(module, "nmea_getinfo.out", "autorun.out")
 
     # getresponse(module, 1)
 
@@ -66,9 +95,16 @@ try:
 
     run_script(module, "canary.out")
     stop_script(module)
+    counter = 0
     while(script_is_running(module)):
         print "Script still running"
-    run_script(module, "nmea_getinfo.out")
+        counter += 1
+        if counter > 20:
+            reset(module)
+    if counter > 20:
+        print("Not running main script as module was reset")
+    else:
+        run_script(module, "nmea_getinfo.out")
 
     read_all(module)
